Move client_public_key validation to avoid circular import error
