+++ concent_api/utils/decorators.py
@@ -15,11 +15,11 @@ from golem_messages.exceptions      import MessageTooOldError
 from golem_messages.exceptions      import TimestampError
 
 from core.validation                import validate_golem_message_signed_with_key
+from core.validation import get_validated_client_public_key_from_client_message
 from core.exceptions                import ConcentInSoftShutdownMode
 from core.exceptions import ConcentFeatureIsNotAvailable
 from core.exceptions                import Http400
 
-from utils.helpers                  import get_validated_client_public_key_from_client_message
 from utils.helpers import join_messages
 from utils.shortcuts                import load_without_public_key
